move guard typer to its own file and data structure
